Simulate Model Matrix and Response Vector

Generates synthetic covariates and response vector from a specified distribution for simulation studies or method validation.

Usage

simulXy(
  n,
  p,
  interc = 0,
  beta,
  family = gaussian(),
  prop = 0.1,
  lim.b = c(-3, 3),
  sigma = 1,
  size = 1,
  rho = 0,
  scale.data = TRUE,
  seed = NULL,
  X = NULL,
  dispersion = 0.1
)

Arguments

n

Integer. Number of observations.

p

Integer. Total number of covariates in the model matrix.

interc

Numeric. Intercept to include in the linear predictor. Default is 0.

beta

Numeric vector of length p. Regression coefficients in the linear predictor.

family

Distribution and link function. Allowed: gaussian(), binomial(), poisson() and , Gamma(). Can be a string, function, or family object.

prop

Numeric in [0,1]. Used only if beta is missing; proportion of non-zero coefficients in p. Default is 0.1.

lim.b

Numeric vector of length 2. Range for coefficients if beta is missing. Default: c(-3, 3).

sigma

Standard deviation of Gaussian response. Default is 1.

size

Integer. Number of trials for binomial response. Default is 1.

rho

Numeric. Correlation coefficient for generating covariates. Used to create AR(1)-type covariance: rho^|i-j|. Default is 0.

scale.data

Logical. Whether to scale columns of the model matrix. Default is TRUE.

seed

Optional. Integer seed for reproducibility.

X

Optional. Custom model matrix. If supplied, it overrides the internally generated X.

dispersion

Dispersion parameter of Gamma response. Default is 0.1.

Value

A list with components:

X

Model matrix of dimension n x p

y

Simulated response vector

beta

True regression coefficients used

eta

Linear predictor

Examples

n <- 100; p <- 100
beta <- c(runif(10, -3, 3), rep(0, p - 10))
sim <- simulXy(n = n, p = p, beta = beta, seed = 1234)
o <- islasso(y ~ ., data = sim$data, family = gaussian())
summary(o, pval = 0.05)
#> 
#> Call:
#> islasso(formula = y ~ ., family = gaussian(), data = sim$data)
#> 
#> Residuals:
#>      Min       1Q   Median       3Q      Max 
#> -2.55104 -0.66441 -0.04485  0.82061  2.73134 
#> 
#>             Estimate Std. Error      Df z value Pr(>|z|)    
#> (Intercept)  -0.1159     0.1268   1.000  -0.914   0.3608    
#> X1           -0.8590     0.1382   1.000  -6.214 5.17e-10 ***
#> X2           -2.5625     0.1459   1.000 -17.565  < 2e-16 ***
#> X3            2.3251     0.1414   1.000  16.443  < 2e-16 ***
#> X4           -2.9219     0.1456   1.000 -20.075  < 2e-16 ***
#> X6           -2.3034     0.1423   1.000 -16.185  < 2e-16 ***
#> X7           -0.5082     0.1367   0.999  -3.719   0.0002 ***
#> X8            1.3835     0.1429   1.000   9.683  < 2e-16 ***
#> X9           -2.0468     0.1449   1.000 -14.125  < 2e-16 ***
#> X10          -0.5986     0.1459   1.000  -4.104 4.06e-05 ***
#> ---
#> Signif. codes:  0 ‘***’ 0.001 ‘**’ 0.01 ‘*’ 0.05 ‘.’ 0.1 ‘ ’ 1
#> 
#> (Dispersion parameter for gaussian family taken to be 1.607661)
#> 
#>     Null deviance: 4334.92  on 99.00  degrees of freedom
#> Residual deviance:  104.83  on 65.21  degrees of freedom
#> AIC: 174.42
#> Lambda: 14.171
#> 
#> Number of Newton-Raphson iterations: 102
#>