General Linear Hypotheses for islasso Models

Tests general linear hypotheses and computes confidence intervals for linear combinations of coefficients from a fitted islasso model.

Usage

# S3 method for class 'islasso'
anova(object, A, b = NULL, ci, ...)

Arguments

object

A fitted model object of class "islasso".

A

Hypothesis specification. Either:

A numeric matrix or vector with each row specifying a linear combination of coefficients,

Or a character vector with symbolic expressions (e.g. "X1 + X2 = 3").

b

Right-hand side vector for the null hypotheses A %*% beta = b. If omitted, defaults to zeros.

ci

Optional 2-column matrix of confidence intervals for coefficients.

...

Currently unused.

Value

An object of class "anova.islasso" containing:

Estimate

Linear combination estimates

SE

Standard errors

Wald

Wald statistics

p-value

Associated p-values

Details

The method tests the null hypothesis \(H_0: A \beta = b\), where \(A\) and \(b\) define a linear constraint on model coefficients.

Symbolic expressions support natural syntax: coefficients may be added/subtracted, constants may be multiplied (e.g. "2 * X1 + 3 * X2 = 7"). Equations with omitted = assume zero on the right-hand side. See examples for syntax flexibility.

See also

Author

Gianluca Sottile gianluca.sottile@unipa.it

Examples

n <- 100; p <- 100
beta <- c(runif(10, -2, 2), rep(0, p - 10))
sim <- simulXy(n = n, p = p, beta = beta, seed = 1, family = gaussian())
fit <- islasso(y ~ . -1, data = sim$data, family = gaussian())

# Test if first 5 variables sum to -7.5
anova(fit, A = c("X1 + X2 + X3 + X4 + X5 = -7.5"))
#> 
#> 	 Simultaneous Tests for General Linear Combination
#> 
#> Fit: islasso(formula = y ~ . - 1, family = gaussian(), data = sim$data)
#> 
#> Linear Hypothesis:
#>                                         Estimate Std. Error Chisq Pr(>Chisq)
#> 1*X1 + 1*X2 + 1*X3 + 1*X4 + 1*X5 = -7.5  11.1047     0.2058  2912     <2e-16
#>                                            
#> 1*X1 + 1*X2 + 1*X3 + 1*X4 + 1*X5 = -7.5 ***
#> ---
#> Signif. codes:  0 ‘***’ 0.001 ‘**’ 0.01 ‘*’ 0.05 ‘.’ 0.1 ‘ ’ 1
#> 

# Test multiple hypotheses
anova(fit, A = c("X1 + X2 + X3 + X4 + X5", "X6 + X7 + X8 + X9 + X10"), b = c(-7.5, 8.75))
#> 
#> 	 Simultaneous Tests for General Linear Combination
#> 
#> Fit: islasso(formula = y ~ . - 1, family = gaussian(), data = sim$data)
#> 
#> Linear Hypothesis:
#>                                          Estimate Std. Error Chisq Pr(>Chisq)
#> 1*X1 + 1*X2 + 1*X3 + 1*X4 + 1*X5 = -7.5   11.1047     0.2058  2912     <2e-16
#> 1*X6 + 1*X7 + 1*X8 + 1*X9 + 1*X10 = 8.75  -9.6003     0.2007  2288     <2e-16
#> Overall                                                       5376     <2e-16
#>                                             
#> 1*X1 + 1*X2 + 1*X3 + 1*X4 + 1*X5 = -7.5  ***
#> 1*X6 + 1*X7 + 1*X8 + 1*X9 + 1*X10 = 8.75 ***
#> Overall                                  ***
#> ---
#> Signif. codes:  0 ‘***’ 0.001 ‘**’ 0.01 ‘*’ 0.05 ‘.’ 0.1 ‘ ’ 1
#> 

# Full diagonal comparison to true coefficients
anova(fit, A = diag(p), b = beta)
#> 
#> 	 Simultaneous Tests for General Linear Combination
#> 
#> Fit: islasso(formula = y ~ . - 1, family = gaussian(), data = sim$data)
#> 
#> Linear Hypothesis:
#>                             Estimate Std. Error  Chisq Pr(>Chisq)  
#> 1*X1 = -0.865338914096355  4.526e-02  9.469e-02  0.228      0.633  
#> 1*X2 = 0.957540181465447  -1.052e-01  9.490e-02  1.229      0.268  
#> 1*X3 = 1.07044170238078   -1.107e-01  9.387e-02  1.391      0.238  
#> 1*X4 = 1.46746669057757   -1.634e-01  9.606e-02  2.893      0.089 .
#> 1*X5 = 1.37849498912692   -6.984e-02  9.732e-02  0.515      0.473  
#> 1*X6 = -1.24812392890453   8.192e-02  9.317e-02  0.773      0.379  
#> 1*X7 = -1.32358047738671  -1.393e-02  9.114e-02  0.023      0.879  
#> 1*X8 = -1.40233596228063   1.491e-01  9.558e-02  2.433      0.119  
#> 1*X9 = 1.66532131284475    2.547e-02  1.005e-01  0.064      0.800  
#> 1*X10 = 1.35188855323941  -1.360e-01  9.968e-02  1.863      0.172  
#> 1*X11 = 0                  3.741e-02  5.452e-02  0.471      0.493  
#> 1*X12 = 0                  4.082e-07  8.355e-03  0.000      1.000  
#> 1*X13 = 0                  2.570e-08  1.705e-03  0.000      1.000  
#> 1*X14 = 0                  4.661e-07  9.593e-03  0.000      1.000  
#> 1*X15 = 0                 -3.779e-03  4.814e-02  0.006      0.937  
#> 1*X16 = 0                 -9.638e-03  4.749e-02  0.041      0.839  
#> 1*X17 = 0                  4.690e-07  9.576e-03  0.000      1.000  
#> 1*X18 = 0                 -8.029e-07  1.623e-02  0.000      1.000  
#> 1*X19 = 0                 -4.643e-02  5.797e-02  0.642      0.423  
#> 1*X20 = 0                  9.580e-06  4.380e-02  0.000      1.000  
#> 1*X21 = 0                  8.822e-02  7.255e-02  1.479      0.224  
#> 1*X22 = 0                 -5.387e-07  1.098e-02  0.000      1.000  
#> 1*X23 = 0                 -1.739e-04  4.104e-02  0.000      0.997  
#> 1*X24 = 0                  6.069e-02  6.177e-02  0.965      0.326  
#> 1*X25 = 0                  3.179e-02  5.352e-02  0.353      0.553  
#> 1*X26 = 0                  1.647e-02  4.981e-02  0.109      0.741  
#> 1*X27 = 0                 -1.717e-07  3.526e-03  0.000      1.000  
#> 1*X28 = 0                 -7.268e-07  1.471e-02  0.000      1.000  
#> 1*X29 = 0                  5.197e-02  5.885e-02  0.780      0.377  
#> 1*X30 = 0                  8.310e-07  1.691e-02  0.000      1.000  
#> 1*X31 = 0                 -7.129e-07  1.454e-02  0.000      1.000  
#> 1*X32 = 0                 -6.172e-07  1.251e-02  0.000      1.000  
#> 1*X33 = 0                 -8.948e-07  1.794e-02  0.000      1.000  
#> 1*X34 = 0                  1.358e-05  4.619e-02  0.000      1.000  
#> 1*X35 = 0                 -8.346e-07  1.700e-02  0.000      1.000  
#> 1*X36 = 0                  2.442e-03  4.310e-02  0.003      0.955  
#> 1*X37 = 0                  4.235e-07  8.703e-03  0.000      1.000  
#> 1*X38 = 0                 -9.916e-03  4.606e-02  0.046      0.830  
#> 1*X39 = 0                 -1.370e-02  4.607e-02  0.088      0.766  
#> 1*X40 = 0                 -1.385e-02  4.323e-02  0.103      0.749  
#> 1*X41 = 0                  5.039e-07  1.030e-02  0.000      1.000  
#> 1*X42 = 0                  2.646e-02  4.941e-02  0.287      0.592  
#> 1*X43 = 0                  6.782e-07  1.388e-02  0.000      1.000  
#> 1*X44 = 0                 -4.573e-02  5.791e-02  0.624      0.430  
#> 1*X45 = 0                 -7.514e-03  4.619e-02  0.026      0.871  
#> 1*X46 = 0                 -5.589e-03  4.593e-02  0.015      0.903  
#> 1*X47 = 0                  4.151e-02  5.502e-02  0.569      0.451  
#> 1*X48 = 0                  1.600e-02  4.740e-02  0.114      0.736  
#> 1*X49 = 0                 -9.059e-07  1.827e-02  0.000      1.000  
#> 1*X50 = 0                  3.641e-02  5.317e-02  0.469      0.493  
#> 1*X51 = 0                  6.857e-07  1.382e-02  0.000      1.000  
#> 1*X52 = 0                  2.850e-07  5.784e-03  0.000      1.000  
#> 1*X53 = 0                  7.639e-08  1.707e-03  0.000      1.000  
#> 1*X54 = 0                  6.030e-02  6.260e-02  0.928      0.335  
#> 1*X55 = 0                 -2.744e-02  5.146e-02  0.284      0.594  
#> 1*X56 = 0                 -6.434e-07  1.308e-02  0.000      1.000  
#> 1*X57 = 0                 -7.862e-07  1.588e-02  0.000      1.000  
#> 1*X58 = 0                  2.846e-02  5.065e-02  0.316      0.574  
#> 1*X59 = 0                  7.632e-03  4.700e-02  0.026      0.871  
#> 1*X60 = 0                  5.287e-07  1.083e-02  0.000      1.000  
#> 1*X61 = 0                  2.142e-02  4.759e-02  0.203      0.653  
#> 1*X62 = 0                  2.570e-02  4.962e-02  0.268      0.604  
#> 1*X63 = 0                  1.444e-02  4.528e-02  0.102      0.750  
#> 1*X64 = 0                 -1.414e-07  2.918e-03  0.000      1.000  
#> 1*X65 = 0                  1.437e-02  4.308e-02  0.111      0.739  
#> 1*X66 = 0                 -4.365e-07  8.849e-03  0.000      1.000  
#> 1*X67 = 0                  4.933e-02  5.871e-02  0.706      0.401  
#> 1*X68 = 0                  8.675e-07  1.745e-02  0.000      1.000  
#> 1*X69 = 0                 -5.880e-07  1.205e-02  0.000      1.000  
#> 1*X70 = 0                  9.456e-07  1.897e-02  0.000      1.000  
#> 1*X71 = 0                  3.522e-03  4.416e-02  0.006      0.936  
#> 1*X72 = 0                  8.105e-07  1.628e-02  0.000      1.000  
#> 1*X73 = 0                 -2.858e-07  5.819e-03  0.000      1.000  
#> 1*X74 = 0                  5.091e-02  6.044e-02  0.710      0.400  
#> 1*X75 = 0                  5.467e-06  4.223e-02  0.000      1.000  
#> 1*X76 = 0                 -1.352e-08  1.705e-03  0.000      1.000  
#> 1*X77 = 0                 -7.959e-07  1.621e-02  0.000      1.000  
#> 1*X78 = 0                 -3.705e-07  7.596e-03  0.000      1.000  
#> 1*X79 = 0                  2.197e-02  4.925e-02  0.199      0.656  
#> 1*X80 = 0                 -3.431e-03  4.403e-02  0.006      0.938  
#> 1*X81 = 0                  6.725e-07  1.381e-02  0.000      1.000  
#> 1*X82 = 0                 -3.746e-07  7.698e-03  0.000      1.000  
#> 1*X83 = 0                 -1.283e-01  8.217e-02  2.436      0.119  
#> 1*X84 = 0                  3.551e-02  5.366e-02  0.438      0.508  
#> 1*X85 = 0                  9.524e-02  7.312e-02  1.697      0.193  
#> 1*X86 = 0                  5.214e-07  1.052e-02  0.000      1.000  
#> 1*X87 = 0                 -1.035e-01  7.530e-02  1.890      0.169  
#> 1*X88 = 0                  5.677e-08  1.705e-03  0.000      1.000  
#> 1*X89 = 0                 -6.283e-02  6.345e-02  0.981      0.322  
#> 1*X90 = 0                  1.482e-07  3.041e-03  0.000      1.000  
#> 1*X91 = 0                 -5.714e-07  1.168e-02  0.000      1.000  
#> 1*X92 = 0                 -4.001e-02  5.675e-02  0.497      0.481  
#> 1*X93 = 0                 -4.762e-07  9.664e-03  0.000      1.000  
#> 1*X94 = 0                  2.652e-07  5.442e-03  0.000      1.000  
#> 1*X95 = 0                  9.546e-03  4.856e-02  0.039      0.844  
#> 1*X96 = 0                 -4.106e-04  4.632e-02  0.000      0.993  
#> 1*X97 = 0                  5.261e-07  1.078e-02  0.000      1.000  
#> 1*X98 = 0                 -4.890e-03  4.125e-02  0.014      0.906  
#> 1*X99 = 0                 -3.266e-07  6.699e-03  0.000      1.000  
#> 1*X100 = 0                 5.958e-07  1.205e-02  0.000      1.000  
#> Overall                                         32.722      1.000  
#> ---
#> Signif. codes:  0 ‘***’ 0.001 ‘**’ 0.01 ‘*’ 0.05 ‘.’ 0.1 ‘ ’ 1
#>